--- pkg/controller/user/disable_user.go
@@ -1,59 +1,0 @@
-// Copyright 2020 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package user
-
-import (
-	"net/http"
-
-	"github.com/google/exposure-notifications-verification-server/pkg/controller"
-	"github.com/google/exposure-notifications-verification-server/pkg/database"
-	"github.com/gorilla/mux"
-)
-
-func (c *Controller) HandleDisableUser() http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		ctx := r.Context()
-		vars := mux.Vars(r)
-
-		session := controller.SessionFromContext(ctx)
-		if session == nil {
-			controller.MissingSession(w, r, c.h)
-			return
-		}
-		flash := controller.Flash(session)
-
-		// Pull the user from the id.
-		user, err := c.db.FindUser(vars["id"])
-		if err != nil {
-			if database.IsNotFound(err) {
-				controller.Unauthorized(w, r, c.h)
-				return
-			}
-
-			controller.InternalError(w, r, c.h, err)
-			return
-		}
-
-		// Disable the user.
-		if err := c.authProvider.DisableUser(ctx, user.Email); err != nil {
-			flash.Error("Failed to disable user: %v", err)
-			controller.Back(w, r, c.h)
-			return
-		}
-
-		flash.Alert("Successfully disabled user to %v", user.Email)
-		controller.Back(w, r, c.h)
-	})
-}
